fix: move folder Vacancies to src/pages
