+++ src/pages/Vacancies/[vacancyId]/vacancy.styles.js
